Issue #20: Moved all of the extraneous files from server/server.go into more specific files. Still part of the server package, but it's easier to grok where things are now.
